+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Fixed
+Skipping region between interrupt addresses and start opcodes
+
 ## [0.1.0] - 2018-04-11
 ### Changed
 License change to MIT license