+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+## [0.2.0] - 2018-05-15
 ### Added
 Parsing of ROM name
 
@@ -29,6 +30,7 @@ URL in project file
 ### Added
 First version parsing bytes up to the scrolling start logo.
 
-[Unreleased]: https://github.com/gernd/gb-dumper/compare/v0.1.0...master
+[Unreleased]: https://github.com/gernd/gb-dumper/compare/v0.2.0...master
+[0.2.0]: https://github.com/gernd/gb-dumper/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/gernd/gb-dumper/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/gernd/gb-dumper/compare/ca6130150d7728ef7190c84a013ebc45b8e383e1...v0.0.1